refactor(host): implement ControlInterfaceServer trait
